File structure changes, and adding CSS link to index.jsp.
The WebContent folder was in an unexpected spot, so I renamed it to WebRoot and moved it up the tree a bit to match what I'm used to seeing (I don't know if this is best practice or not, still too new to this).

Added link to a CSS file on index.jsp, and created the linked CSS file.